+++ plugins/gui/src/selection_details_widget/module_details_widget/module_ports_tree.cpp
@@ -6,15 +6,7 @@
 #include "gui/input_dialog/pingroup_selector_dialog.h"
 #include "gui/python/py_code_provider.h"
 #include "gui/selection_details_widget/module_details_widget/port_tree_model.h"
-#include "gui/user_action/action_add_items_to_object.h"
-#include "gui/user_action/action_create_object.h"
-#include "gui/user_action/action_delete_object.h"
-#include "gui/user_action/action_remove_items_from_object.h"
-#include "gui/user_action/action_rename_object.h"
-#include "gui/user_action/action_set_object_type.h"
 #include "gui/user_action/action_pingroup.h"
-#include "gui/user_action/user_action_compound.h"
-#include "hal_core/netlist/gate_library/enums/pin_direction.h"
 #include "hal_core/netlist/gate_library/enums/pin_type.h"
 #include "hal_core/utilities/enums.h"
 
@@ -142,20 +134,42 @@ namespace hal
 
         if (clickedItem->itemType() == PortTreeItem::Group)    //group specific context, own helper function? (returns at the end)
         {
-            menu.addAction("Change name", [name, modId, itemId]() {
+            menu.addAction("Change name", [itemId, mod, name]() {
+                PinGroup<ModulePin>*  pg = mod->get_pin_group_by_id(itemId);
+                if (pg == nullptr) return;
+
                 InputDialog ipd("Change pin group name", "New group name", name);
                 if (ipd.exec() == QDialog::Accepted)
                 {
                     if (ipd.textValue().isEmpty())
                         return;
-                    auto* group = gNetlist->get_module_by_id(modId)->get_pin_group_by_id(itemId);
-                    if (group != nullptr)
-                    {
-                        ActionPingroup* act = new ActionPingroup(PinActionType::GroupRename,itemId,ipd.textValue());
-                        act->setObject(UserActionObject(modId, UserActionObjectType::Module));
-                        act->exec();
-                    }
+
+                    ActionPingroup* act = new ActionPingroup(PinActionType::GroupRename,itemId,ipd.textValue());
+                    act->setObject(UserActionObject(mod->get_id(), UserActionObjectType::Module));
+                    act->exec();
                 }
+            });
+
+            menu.addAction("Change type", [itemId, mod, name]() {
+                PinGroup<ModulePin>* pg = mod->get_pin_group_by_id(itemId);
+                if (pg == nullptr) return;
+
+                QStringList types;
+                for (auto const& [k, v] : EnumStrings<PinType>::data)
+                    types << QString::fromStdString(v);
+
+                ComboboxDialog cbd("Change type", "Select type for pin group " + name, types);
+
+                if (cbd.exec() == QDialog::Accepted)
+                {
+                    PinType ptype = enum_from_string<PinType>(cbd.textValue().toStdString(),PinType::none);
+                    ActionPingroup* act = ActionPingroup::changePinGroupType(mod,itemId,(int)ptype);
+                    if (act) act->exec();
+                }
+            });
+            menu.addAction("Toggle ascending/descending", [itemId, mod](){
+                ActionPingroup* act = ActionPingroup::toggleAscendingGroup(mod, itemId);
+                if (act) act->exec();
             });
             menu.addAction("Delete pin group", [itemId, mod]() {
                 auto* pinGroup = mod->get_pin_group_by_id(itemId);