+++ plugins/gui/src/selection_details_widget/module_details_widget/port_tree_model.cpp
@@ -2,14 +2,10 @@
 
 #include "gui/basic_tree_model/base_tree_item.h"
 #include "gui/gui_globals.h"
-#include "gui/input_dialog/input_dialog.h"
 #include "gui/python/py_code_provider.h"
 #include "gui/user_action/action_pingroup.h"
-#include "gui/user_action/action_remove_items_from_object.h"
-#include "gui/user_action/user_action_compound.h"
 #include "hal_core/netlist/endpoint.h"
 #include "hal_core/netlist/gate.h"
-#include "hal_core/netlist/gate_library/gate_type.h"
 #include "hal_core/netlist/module.h"
 #include "hal_core/netlist/net.h"
 #include "hal_core/utilities/enums.h"
@@ -209,8 +205,17 @@ namespace hal
             }
             else
             {
+                PortTreeItem* pitem = dynamic_cast<PortTreeItem*>(parentItem);
+                if (pitem && pitem->itemType() == PortTreeItem::Pin)
+                {
+// debug pingroup                qDebug() << "pin was dropped on a pin...";
+                    PortTreeItem* ppitem = static_cast<PortTreeItem*>(pitem->getParent());
+                    row = getIndexFromItem(pitem).row();
+                    dndPinBetweenPin(droppedItem, ppitem, row);
+                }
+                else
 // debug pingroup                qDebug() << "pin was dropped on a group...";
-                dndPinOnGroup(droppedItem, parentItem);
+                    dndPinOnGroup(droppedItem, parentItem);
             }
         }
 
@@ -256,8 +261,10 @@ namespace hal
             // perhaps check here that a pin can only be dropped between groups if its own group has size > 1?
             // otherwise it does not make much sense...perhaps change this check
             auto item = mIdToPinItem[id];
-            if((!parentItem && item->getParent()->getChildCount() == 1) || (item->getParent() == parentItem && row == -1) || item == parentItem
-                || (parentItem && (parentItem->itemType() == PortTreeItem::Pin)))
+            if((!parentItem && item->getParent()->getChildCount() == 1)
+                || (item->getParent() == parentItem && row == -1)
+                || item == parentItem )
+               // || (parentItem && (parentItem->itemType() == PortTreeItem::Pin)))
                 return false;
             // case if one wants to drop between pins in same group, check if its not adjacent row (other cases are handled on case above
             if(item->getParent() == parentItem)
@@ -515,6 +522,7 @@ namespace hal
                                    netName);
             mIdToPinItem.insert(ptiPin->id(), ptiPin);
             insertItem(ptiPin, ptiGroup, pinRow);
+            updateGroupIndex(ptiGroup);
             break;
         }
         case PinEvent::PinReorder:
@@ -526,6 +534,7 @@ namespace hal
             }
             removeItem(ptiPin);
             insertItem(ptiPin, ptiGroup, pinRow);
+            updateGroupIndex(ptiGroup);
             break;
         }
         case PinEvent::PinAssignToGroup:
@@ -537,6 +546,7 @@ namespace hal
             }
             removeItem(ptiPin);
             insertItem(ptiPin, ptiGroup, pinRow);
+            updateGroupIndex(ptiGroup);
             break;
         }
         case PinEvent::PinRename:
@@ -641,6 +651,23 @@ namespace hal
         act->exec();
     }
 
+    void ModulePinsTreeModel::updateGroupIndex(PortTreeItem* groupItem)
+    {
+        PinGroup<ModulePin>* pg = mModule->get_pin_group_by_id(groupItem->id());
+        Q_ASSERT(pg);
+        for (ModulePin* pin : pg->get_pins())
+        {
+            int inx = pg->get_index(pin).get();
+            PortTreeItem* pinItem = mIdToPinItem.value(pin->get_id());
+            Q_ASSERT(pinItem);
+            pinItem->setIndex(inx);
+        }
+        QModelIndex pi = getIndexFromItem(groupItem);
+        QModelIndex i0 = index(0,0,pi);
+        QModelIndex i1 = index(groupItem->getChildCount()-1,4,pi);
+        Q_EMIT dataChanged(i0,i1);
+    }
+
     void ModulePinsTreeModel::insertItem(PortTreeItem* item, BaseTreeItem* parent, int index)
     {
         // fun fact: if an item is inserted above an item that is expanded, the tree collapses all indeces
@@ -651,6 +678,7 @@ namespace hal
         item->itemType() == PortTreeItem::Pin ? mIdToPinItem.insert(item->id(), item) : mIdToGroupItem.insert(item->id(), item);
         //mIdToPinItem.insert(getIdOfItem(item), item);
     }
+
     void ModulePinsTreeModel::removeItem(PortTreeItem* item)
     {
         beginRemoveRows(parent(getIndexFromItem(item)), item->getOwnRow(), item->getOwnRow());