+++ CHANGELOG.md
@@ -14,7 +14,12 @@ All notable changes to this project will be documented in this file.
   * added build support for RedHat Enterprise Linux
   * fixed `netlist_preprocessing` build dependencies
 
-* added column for pin index in pin tree shown in selection details
+* pin groups
+  * added getter for lowest and highest index of pin group
+  * added column for pin index in pin tree shown in selection details
+  * added context menu entry to toggle between ascending and descending
+  * changed drag&drop behavior, allowing to drop pin on pin when ordering
+   
 * fixed availability of "save as" so that does not required modifications to be enabaled
 * added feature to unzip and open hal project by dropping zipped file on welcome screen
 * added `get_shortest_path` overload to find shortest path from gate to module